move assembleFeatureLabelCols from utils to shards
